pyproject.toml
[tool.poetry]
name = "cloudflare-b2"
version = "0.1.0"
description = ""
authors = ["Backblaze"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
